requirements.txt
aiohttp>=3.9.0
scipy>=1.11.0
python-dotenv>=1.0.0
